Drop harfbuzz glib dependency
This is not strictly required - if not present, Harfbuzz will use its built-in Unicode character table. This ensures that the library can always load, even if the system does not have glib installed.
The built-in Unicode table was previously included anyway, so this doesn't come with an additional size cost.